Finish 03 flex exercise (with solution corrections)
I started off with the right idea that the right side needed to be in
its own <div> in the html. I didn't realize that I would have split
both the left and right side up, giving the css the same properties. I
also complicate things a bit. This is definitely something that will
just take me more time and practice to get used to